Finish Library class development and Refactor some code
Relocated the Logging-related methods into the Library class, because
it's the class responsible for controlling and managing the inclusion and
removal of Books from the Library

Created methods to save Book-related info on a .csv file, although it
can be polished a little, it enables persistence across different
runs of the application.

Made some miscellaneous changes across the whole project to work with
the "centralization" of control to the Library class, while keeping the
overall funcionalities the same.