Terminando implementação da Classe Library (#5)
* Create the Library class as a Singleton

Created the Library.java file and wrote some logic to fit this class
into the Singleton design pattern

* Implementar Livro e Categoria

* Adicionar Livro e Categoria base, e main para testar

* Deletar out/production/projeto-java directory

* Deletar pasta quarkus

* Atualizar o Book.java

- permits eBook
- atributos alterados para final(com remoção dos setters)

---------

Co-authored-by: Jean Fellipe <jeanfellipesilva93@gmail.com>
Co-authored-by: Jean Fellipe <105667153+codingmage@users.noreply.github.com>
Co-authored-by: AlexMipi <107962033+AlexMipi@users.noreply.github.com>

* Classes Author e eBook (#3)

* Alterações feitas em Autor e eBook

* Construtor para Autor

* Interface Logger, Método updateLog para PhysicalBook e Bugfixes

Bugfixes:
Construtor de PhysicalBook atualizado
Métodos da interface Lendable corrigido

* Create the Library class as a Singleton

Created the Library.java file and wrote some logic to fit this class
into the Singleton design pattern

* Start Library class implementation and Create Validator class

In order to reduce the complexity of the Main class, refactored
the book managing methods into the Library class, so that the Main
calls the Methods from Library as needed

Created Validator class in order to check if the ISBN13 input is
actually valid, based on the checking digit of the ISBN13 String

* Implementar Livro e Categoria

* Adicionar Livro e Categoria base, e main para testar

* Deletar out/production/projeto-java directory

* Deletar pasta quarkus

* Atualizar o Book.java

- permits eBook
- atributos alterados para final(com remoção dos setters)

---------

Co-authored-by: Jean Fellipe <jeanfellipesilva93@gmail.com>
Co-authored-by: Jean Fellipe <105667153+codingmage@users.noreply.github.com>
Co-authored-by: AlexMipi <107962033+AlexMipi@users.noreply.github.com>

* Create the Library class as a Singleton

Created the Library.java file and wrote some logic to fit this class
into the Singleton design pattern

* Start Library class implementation and Create Validator class

In order to reduce the complexity of the Main class, refactored
the book managing methods into the Library class, so that the Main
calls the Methods from Library as needed

Created Validator class in order to check if the ISBN13 input is
actually valid, based on the checking digit of the ISBN13 String

* Polish Library class implementation

Abstracted the menu shown when running the application to a method
in the Library class, that can be called upon necessity

Changed some things in the Main class to show the created book details
when the process to create one has finished, so that the user can
see the details sent to the server when finishing the process

* Finish Library class development and Refactor some code

Relocated the Logging-related methods into the Library class, because
it's the class responsible for controlling and managing the inclusion and
removal of Books from the Library

Created methods to save Book-related info on a .csv file, although it
can be polished a little, it enables persistence across different
runs of the application.

Made some miscellaneous changes across the whole project to work with
the "centralization" of control to the Library class, while keeping the
overall funcionalities the same.

* add .gitignore

* Include validateISBN13 into Book class

---------

Co-authored-by: Jean Fellipe <jeanfellipesilva93@gmail.com>
Co-authored-by: Jean Fellipe <105667153+codingmage@users.noreply.github.com>
Co-authored-by: AlexMipi <107962033+AlexMipi@users.noreply.github.com>
Co-authored-by: jgribeiro83 <joaogabrielrr@yahoo.com.br>